Move cwd to app header
