+++ packages/cli/src/ui/utils/isNarrowWidth.ts
@@ -5,5 +5,5 @@
  */
 
 export function isNarrowWidth(width: number): boolean {
-  return width < 80;
+  return width < 70;
 }
